biddie.io/biddie_backend/requirements.txt
boto==2.36.0
Cheetah==2.4.4
Flask==0.10.1
Flask-Bcrypt==0.6.0
Flask-PyMongo==0.3.0
Jinja2==2.7.3
Markdown==2.5.2
MarkupSafe==0.23
PyYAML==3.10
Werkzeug==0.9.6
enum==0.4.4
gnureadline==6.3.3
ipdb==0.8
ipython==2.3.1
itsdangerous==0.24
py-bcrypt==0.4
pymongo==2.7.2
pysqlite==2.6.3
pyzipcode==0.4
requests==0.13.2
stripe==1.19.1
urllib3==1.7.1
wsgiref==0.1.2

